Extract DesktopNavbar from Navbar
